Removing _doatexits()
Fixing malloc (APEX, not native).
It warns like a hell, but as Aki said last summer,
casts are only a mask for a dumb compiler and very dangerous.

Now it's ok.

Added Harvey to ap/plan9/uname.c

+++ lib/ap/plan9/malloc.c
@@ -37,7 +37,7 @@ struct Arena
 };
 static Arena arena;
 
-#define datoff		((int64_t)((Bucket*)0)->data)
+#define datoff		((intptr_t)((Bucket*)0)->data)
 #define nil		((void*)0)
 
 extern	void	*sbrk(unsigned long);
@@ -45,7 +45,7 @@ extern	void	*sbrk(unsigned long);
 void*
 malloc(size_t size)
 {
-	uint64_t next;
+	uint next;
 	int pow, n;
 	Bucket *bp, *nbp;
 
@@ -74,15 +74,15 @@ good:
 	if(pow < CUTOFF) {
 		n = (CUTOFF-pow)+2;
 		bp = sbrk(size*n);
-		if((int64_t)bp == -1)
+		if((intptr_t)bp == -1)
 			return nil;
 
-		next = (uint64_t)bp+size;
-		nbp = (Bucket*)(uint64_t)next;
+		next = (uintptr_t)bp+size;
+		nbp = (Bucket*)next;
 		arena.btab[pow] = nbp;
 		for(n -= 2; n; n--) {
-			next = (uint64_t)nbp+size;
-			nbp->next = (Bucket*)(uint64_t)next;
+			next = (uintptr_t)nbp+size;
+			nbp->next = (Bucket*)next;
 			nbp->size = pow;
 			nbp = nbp->next;
 		}
@@ -90,7 +90,7 @@ good:
 	}
 	else {
 		bp = sbrk(size);
-		if((int64_t)bp == -1)
+		if((intptr_t)bp == -1)
 			return nil;
 	}
 
@@ -109,7 +109,7 @@ free(void *ptr)
 		return;
 
 	/* Find the start of the structure */
-	bp = (Bucket*)((uint64_t)ptr - datoff);
+	bp = (Bucket*)((uintptr_t)ptr - datoff);
 
 	if(bp->magic != MAGIC)
 		abort();
@@ -131,7 +131,7 @@ realloc(void *ptr, size_t n)
 		return malloc(n);
 
 	/* Find the start of the structure */
-	bp = (Bucket*)((uint64_t)ptr - datoff);
+	bp = (Bucket*)((uintptr_t)ptr - datoff);
 
 	if(bp->magic != MAGIC)
 		abort();
